Detect the language that the abstract or other fields are written in

Currently this only works for files that were generated while Citavi was set to "English" so that column names are "Short Title" etc.

detect_language(
  CitDat,
  fieldsToDetectIn = c("Abstract"),
  wantedLanguage = c("english")
)

Arguments

CitDat
A dataframe/tibble possibly returned by read_Citavi_xlsx.

fieldsToDetectIn
Character vector with names of fields whose text language should be detected. Default is c("Abstract"). When multiple fields are given (e.g. c("Abstract", "Title")), they are combined into a single string whose language is then detected.

wantedLanguage
Character vector with names of languages that are desired. Default is c("english"). If not set to NULL, a new column det_lang_wanted is created, which is TRUE if the detected language in det_lang is a wanted language.

Value

A tibble containing at least one additional column: det_lang.

Examples

path <- example_xlsx("3dupsin5refs.xlsx")
read_Citavi_xlsx(path) %>%
   detect_language()

#> # A tibble: 5 x 21
#>   ID     `Short title`  Title Year  Author Categories Groups Abstract `DOI name`
#>   <chr>  <chr>          <chr> <chr> <chr>  <chr>      <chr>  <chr>    <chr>     
#> 1 7e04d~ Schmidt, Hart~ Esti~ 2019  Schmi~ 1 catA     Googl~ Broad-s~ 10.2135/c~
#> 2 2481c~ Schmidt, Hart~ Heri~ 2019  Schmi~ 1 catA     Googl~ In plan~ 10.1534/g~
#> 3 db3ab~ Schmidt, Hart~ Heri~ 2019  Schmi~ 2 catB     PubMed In plan~ NA        
#> 4 ba575~ Schmidt, Hart~ Hrit~ 2019  Schmi~ 1 catA     TypoDB In plan~ 10.1534/g~
#> 5 fa402~ Schmidt, Möhr~ More~ 2018  Schmi~ 2 catB     Googl~ Traditi~ 10.2135/c~
#> # ... with 12 more variables: PubMed ID <chr>, Online address <chr>,
#> #   Periodical <chr>, Volume <chr>, Number <chr>, Page range <chr>,
#> #   Locations <chr>, has_attachment <lgl>, red_flag <lgl>, blue_circle <lgl>,
#> #   det_lang <chr>, det_lang_wanted <lgl>