Identify obvious duplicates based on title and year

Currently this only works for files that were generated while Citavi was set to "English" so that column names are "Short Title" etc.

find_obvious_dups(CitDat, dupInfoAfterID = TRUE, preferDupsWithPDF = TRUE)

Arguments

CitDat
A dataframe/tibble returned by read_Citavi_xlsx. The following columns must be present: ID, Title, Year.

dupInfoAfterID
If TRUE (default), the newly created columns clean_title, clean_title_id, has_obv_dup and obv_dup_id are moved right next to the ID column.

preferDupsWithPDF
If TRUE (default), obvious duplicates are sorted by their info in columns has_attachment and/or Locations (given they are present in the dataset). After sorting, duplicates with the most occurences of ".pdf" in Locations and a TRUE in has_attachment are first and will thus be chosen as dup_01.

Value

A tibble containing three additional columns: clean_title_id, has_obv_dup and obv_dup_id.

Examples

path <- example_xlsx("3dupsin5refs.xlsx")
read_Citavi_xlsx(path) %>%
   find_obvious_dups()

#> # A tibble: 5 x 23
#>   ID     clean_title  clean_title_id has_obv_dup obv_dup_id `Short title`  Title
#>   <chr>  <chr>        <chr>          <lgl>       <chr>      <chr>          <chr>
#> 1 7e04d~ estimating_~ ct_01          FALSE       dup_01     Schmidt, Hart~ Esti~
#> 2 2481c~ heritabilit~ ct_02          TRUE        dup_01     Schmidt, Hart~ Heri~
#> 3 db3ab~ heritabilit~ ct_02          TRUE        dup_02     Schmidt, Hart~ Heri~
#> 4 ba575~ hritability~ ct_03          FALSE       dup_01     Schmidt, Hart~ Hrit~
#> 5 fa402~ more_larger~ ct_04          FALSE       dup_01     Schmidt, Möhr~ More~
#> # ... with 16 more variables: Year <chr>, Author <chr>, Categories <chr>,
#> #   Groups <chr>, Abstract <chr>, DOI name <chr>, PubMed ID <chr>,
#> #   Online address <chr>, Periodical <chr>, Volume <chr>, Number <chr>,
#> #   Page range <chr>, Locations <chr>, has_attachment <lgl>, red_flag <lgl>,
#> #   blue_circle <lgl>