Import Excel files created via Citavi Export

Import Excel files created via Citavi Export

read_Citavi_xlsx(
  path = NULL,
  keepMarksCols = TRUE,
  useYearDerived = TRUE,
  setSuggestedColOrder = TRUE,
  ...
)

Arguments

path
Path to the xls/xlsx file created with Citavi via export to Excel.

keepMarksCols
If TRUE (default) it will keep and rename the first three nameless columns of the imported excel file. These columns are automatically created by the Citavi export and contain information about the attachment/paper-clip marker, the red flag marker and the blue circle marker. If kept, the columns are renamed has_attachment, red_flag and blue_circle. If FALSE, these three columns are deleted.

useYearDerived
If TRUE (default) the special column "Year derived" (DE: "Jahr ermittelt") is renamed to "Year" (DE:"Jahr") and thus replaces the original basic column "Year" (DE: "Jahr") which may have also been created via the Citavi export.

setSuggestedColOrder
If TRUE (default) columns a reordered in a suggested order.

...
Other arguments passed to the readxl::read_xlsx() function.

Value

A tibble containing the information of the xls/xlsx file created with Citavi via export to Excel.

Details

Currently this only works for files that were generated while Citavi was set to "English" or "German" so that column names are "Short title" or "Kurztitel" etc.

Examples

path <- example_xlsx("3dupsin5refs.xlsx") # use this package's example xlsx file
read_Citavi_xlsx(path)

#> # A tibble: 5 x 19
#>   ID     `Short title`  Title Year  Author Categories Groups Abstract `DOI name`
#>   <chr>  <chr>          <chr> <chr> <chr>  <chr>      <chr>  <chr>    <chr>     
#> 1 7e04d~ Schmidt, Hart~ Esti~ 2019  Schmi~ 1 catA     Googl~ Broad-s~ 10.2135/c~
#> 2 2481c~ Schmidt, Hart~ Heri~ 2019  Schmi~ 1 catA     Googl~ In plan~ 10.1534/g~
#> 3 db3ab~ Schmidt, Hart~ Heri~ 2019  Schmi~ 2 catB     PubMed In plan~ NA        
#> 4 ba575~ Schmidt, Hart~ Hrit~ 2019  Schmi~ 1 catA     TypoDB In plan~ 10.1534/g~
#> 5 fa402~ Schmidt, Möhr~ More~ 2018  Schmi~ 2 catB     Googl~ Traditi~ 10.2135/c~
#> # ... with 10 more variables: PubMed ID <chr>, Online address <chr>,
#> #   Periodical <chr>, Volume <chr>, Number <chr>, Page range <chr>,
#> #   Locations <chr>, has_attachment <lgl>, red_flag <lgl>, blue_circle <lgl>

if (FALSE) {
CitDat <- read_Citavi_xlsx("data/yourCitaviExport.xlsx")
}